test: search bar
